+++ Anima/diffusion/tractography/animaBaseProbabilisticTractographyImageFilter.hxx
@@ -83,7 +83,7 @@ BaseProbabilisticTractographyImageFilter <TInputModelImageType>
     if (stepData == 0)
         stepData = 1;
 
-    unsigned int numSteps = std::floor(m_PointsToProcess.size() / (float)stepData);
+    unsigned int numSteps = std::floor(m_PointsToProcess.size() / (double)stepData);
     if (m_PointsToProcess.size() % stepData != 0)
         numSteps++;
